Chore: Review Contracts (#7)
* chore: add contracts back for review

* refactor: improve simulate deploy script

Wait until anvil is listening to the proper port instead of a hardcoded delay

* feat: add `jarInputConduit` to the initial supported components

* docs: update component docs and add usage instructions

* fix: broken test after adding `jarInputConduit`

* chore: setup

* chore: add remove deal fn, adjust auth

* chore: add missing interface

* chore: replace OZ EnumerableSet with an exact copy of the last version

* refactor: remove custom errors in favor of error messages

* refactor: re-introduce deal removal with additional checks

* feat: add component iteration

* Update src/RwaRegistry.t.sol License


* refactor: remove `uint8` in favor of `uint256` in all public-facing methods

* test: check for failure when `variant` is invalid

---------

+++ script/DeployRwaRegistry.s.sol
@@ -57,7 +57,7 @@ contract DeployRwaRegistry is Script {
         addrs[2] = chainlog.getAddress("RWA008_A_INPUT_CONDUIT");
         addrs[3] = rwaLiquidationOracle;
 
-        uint8[] memory variants = new uint8[](4);
+        uint256[] memory variants = new uint256[](4);
         variants[0] = 2; // RwaUrn2
         variants[1] = 2; // RwaOutputConduit2
         variants[2] = 2; // RwaInputConduit2
@@ -79,7 +79,7 @@ contract DeployRwaRegistry is Script {
         addrs[2] = chainlog.getAddress("RWA009_A_OUTPUT_CONDUIT");
         addrs[3] = rwaLiquidationOracle;
 
-        uint8[] memory variants = new uint8[](4);
+        uint256[] memory variants = new uint256[](4);
         variants[0] = 2; // RwaUrn2
         variants[1] = 1; // RwaJar
         variants[2] = type(uint8).max; // Regular address
